chore: use ipfs-unixfs final release instead of fork
BREAKING CHANGE: We rely now on an extended version of the new ipfs blockstore interface to be compatible with the new ipfs-unixfs. Previously unixfs was using a simple block API from ipld which returned blocks everywhere, while now put returns a void Promise and get returns the bytes instead of the Block. Destroy was also renamed to close to be the same as the interface. The only addition in the interface is the blocks async iterator function. All the blocktore implementations extend BlockstoreAdapter to have all the API expected by the interface.